Microcontroller-Codes/Dark-Knight-Programmer/requirements.txt
bleak>=0.20.0
pyserial>=3.5